chore(tools): remove deprecated run_query - query is the canonical SQL tool
The qb-based query tool (PR #140) has been live for the soak period
and exercised across DISTINCT / OFFSET / COUNT(DISTINCT) / EXISTS
patterns. Dropping run_query in this PR closes the 3-repo deprecation
window the v0.1+v0.2 PRs opened.

Changes:
- Delete jarvis/tools/run_query.py and jarvis/tests/test_run_query.py
- Drop "run_query" from tools/registry.py and tests/test_registry.py
- Inline _load_doctype_allowlist into query.py (the only thing query
  was importing from run_query). The Jarvis Settings field name stays
  as run_query_doctype_allowlist to avoid a migration; one operator
  knob continues to gate the SQL tool surface.
- Retarget the dangling "aggregate via run_query" hint in get_list.py
  to point at query.

The plugin (RunQuerySchema + tool descriptor) and persona (frappe-
core SKILL.md "use run_query only when..." caveat) get matched
removal PRs in their own repos.
