+++ jarvis/tools/registry.py
@@ -25,7 +25,6 @@ _TOOL_NAMES: tuple[str, ...] = (
     "get_doc",
     "get_list",
     "run_report",
-    "run_query",
     "query",
     "update_doc",
     "create_doc",
